requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
streamlit
Pillow
